refactor: move locale toggle to separate component
