+++ src/app/core/components/locale-toggler/locale-toggler.component.html
@@ -1,0 +1,7 @@
+<button mat-button class="header-locales">
+  <mat-select *ngIf="selectedLocale" [(ngModel)]="selectedLocale" (selectionChange)="changeLocale($event)">
+    <mat-option *ngFor="let locale of locales" [value]="locale">
+      {{ locale | translate }}
+    </mat-option>
+  </mat-select>
+</button>
